requirements.txt
altair==4.2.0
bsddb3==6.2.9
Flask==2.0.2
fuzzywuzzy==0.18.0
gensim==4.0.1
graphviz==0.19.1
icecream==2.1.1
itemadapter==0.4.0
kglab==0.4.3
matplotlib==3.5.1
mkrefs==0.2.0
networkx==2.6.3
nltk==3.6.5
numpy==1.21.2
pandas==1.3.5
path.py==12.5.0
pymongo==4.0.1
pyrouge==0.1.3
pyrsistent==0.18.0
pytest==7.0.0
PyYAML==6.0
rdflib==6.1.1
regex==2021.8.3
rouge_score==0.0.4
scipy==1.7.3
scrapy==2.5.1
selenium==4.1.0
setuptools==58.0.4
spacy==3.2.1
torch==1.10.1
transformers==4.11.3
Unidecode==1.3.2
urlextract==1.5.0
